STLink.Commands

Description

All the raw commands that an STLink dongle understands can be found here, though high-level wrappers for them are elsewhere.

Documentation

getVersion :: STLink ResponseVersion

Get the version of an STLink dongle.

data ResponseVersion

The STLink version reported by the dongle. The response also contians the VID and PID of the USB interface, but this is ignored since it is redundant.

Constructors

ResponseVersion

Fields

versionStlink :: Int

The STLink protocol version this dongle understands.

versionJtag :: Int

Which JTAG revision the dongle is capable of.

versionSwim :: Int

Purpose unclear.

getCurrentMode :: STLink STLinkMode

Figure out what major mode the STLink dongle is operating in. (Debug or otherwise.)

data STLinkMode

Major mode for STLink, rather than the minor debug mode (DebugMode).

Constructors

ModeDFU

ModeMass

ModeDebug

ModeSWIM

ModeBootloader

getCurrentVoltage :: STLink Float

Get the voltage on the output of the STLink regulator. (Should be ~3.3V).

enterMode :: DebugMode -> STLink ()

Leave either DFU or Mass mode and enter one of the three debug modes.

data DebugMode

Which debugging interface the dongle with interface with. Prefer SWD when possible.

Constructors

DebugJTAG

DebugSWD

DebugSWIM

getStatus :: STLink DebugStatus

Get the current state of the CPU. You must be in SWD or JTAG mode to issue this command.

data DebugStatus

Determines if the CPU is running or halted (through debug commands).

haltCore :: STLink ()

Stop execution on the microcontroller.

runCore :: STLink ()

Continue execution on the microcontroller.

stepCore :: STLink ()

Step one instruction on the microcontroller.

resetCore :: STLink ()

Reset the microcontroller.

readReg :: Reg -> STLink Word32

Read the specified real register. (Debug registers are not accessible through this command).

writeReg :: Reg -> Word32 -> STLink ()

Write a Word32 to the specified Reg.

readMem :: Word32 -> Word16 -> STLink ByteString

Peeks l bytes of memory at address a.

A maximum of 64 bytes can be read in one command.

writeMem :: Word32 -> ByteString -> STLink ()

Writes a ByteString into the location in memory specified by a.

A maximum of 64 bytes can be written in one command.