STLink.Detection

Description

This module can detect attached STLink dongles, and has some fnuctions for pretty-printing a table of detected devices, as well as prompting the user to select one if more than one is found.

Documentation

pickBoard :: IO (Maybe Device)

Ask the user to select one of the available board, or fail if none are available.